Fixed problems caused by time zones
Some of the date tests were failing when I ran them here in
California. I can imagine they pass just fine in Austria, as it
is only 1 hour away from UTC, so there is very little difference.
But here in California it is 8 hours behind UTC, so the JS engine
was offsetting by 8 hours automatically, and date was coming out as
the previous day. I fixed the test by making the date being tested
unambiguous. That way, the expected output should be the same no
matter which time zone the test is run.