Implemented the trans and blocktrans tags
Notes:

- Implemented unit tests for both tags which show how they work
in normal situations, and how they throw exceptions in the right
situations

- Added Loader.findResourcesDir() so that callers can find the
first resources directory along the same path where the loaders
load the templates. This way, the translations and the templates
can go together, where-ever they are located.

- Modified the environment to call findResourcesDir once.

- This code depends on ilib for translation. Debating on whether
or not this should be the case. On the one hand, more i18n should
happen in the filters, as the support is lacking there. Ilib can
help with that. On the other, it would be cleaner if reinhardt
could stand alone...