Now does not depend on ilib for translation
If you create a Resources class with a translate method, the code
will use that to do translation. Also, you can pass a resBundleFactory
method to the environment as you construct it to create new bundles
as necessary. Of course, all of it can be implemented using ilib,
but it is up to the app to choose which translation solution they
prefer as long as they wrap it in the interface mentioned above.